Python/langchain/pyproject.toml
[tool.poetry]
name = "python-openai-autoblocks-example"
version = "0.0.0"
description = "Automatic tracing of LangChain"
authors = ["Autoblocks Engineering <engineering@autoblocks.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
autoblocksai = "^0.0.11"
python-dotenv = "^1.0.0"
langchain = "^0.0.319"
openai = "^0.28.1"
